% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rts.R
\name{find_rts_region}
\alias{find_rts_region}
\title{Finds Region Specific Reproductive Numbers}
\usage{
find_rts_region(data, reg_mat, si_shape = 1.47, si_rate = 0.04529,
  k = 100)
}
\arguments{
\item{data}{Dataframe to be used for estimating reproductive numbers. Must
contain 'cases' column and numerical 'region' column.}

\item{reg_mat}{Matrix (4 x 4) representing interactions between regions.}

\item{si_shape}{Shape parameter for serial interval.}

\item{si_rate}{Rate parameter for serial interval.}

\item{k}{Maximum length of serial interval.}
}
\value{
List of effective reproductive numbers for each regions and
  corresponding week numbers.
}
\description{
Using data and specified serial interval, finds the effective reproductive
  for each region while allowing for interaction between regions. The
  default serial interval uses an estimate for the United States on the
  weekly time scale.
}
