% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rts.R
\name{find_rts_strata}
\alias{find_rts_strata}
\title{Finds Region Specific Reproductive Numbers}
\usage{
find_rts_strata(data, num_strata = 4, strata_mat = diag(4),
  si_shape = 1.47, si_rate = 0.04529, k = 100)
}
\arguments{
\item{data}{Dataframe to be used for estimating reproductive numbers. Must
contain 'cases' column and numerical 'region' column.}

\item{num_strata}{Integer for the number of strata that the data is being
split by. Default is 4.}

\item{strata_mat}{Matrix representing interactions between strata.}

\item{si_shape}{Shape parameter for serial interval.}

\item{si_rate}{Rate parameter for serial interval.}

\item{k}{Maximum length of serial interval.}
}
\value{
List of effective reproductive numbers for each stratum and
  corresponding week numbers. For analysis, these reproductive numbers must
  be seperated.
}
\description{
Using data and specified serial interval, finds the effective reproductive
  for each region while allowing for interaction between regions. The
  default serial interval uses an estimate for the United States on the
  weekly time scale.
}
