% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rts_setup.R
\name{get_ps}
\alias{get_ps}
\title{Find Discrete Serial Interval}
\usage{
get_ps(si_shape, si_rate, k = 100)
}
\arguments{
\item{si_shape}{A float, shape parameter for gamma distribution.}

\item{si_rate}{A float, rate parameter for gamma distribution.}

\item{k}{An integer, maximum length of gamma distribution}
}
\value{
List with values for discrete gamma distribution with length
  \code{k}.
}
\description{
Gets discrete probability distribution for serial interval using given
  shape and rate parameters.
}
\examples{
getPs(si_shape=1.47, si_rate=0.04529, k=100)
}
