% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rts_setup.R
\name{get_rts}
\alias{get_rts}
\title{Finds Reproductive Numbers}
\usage{
get_rts(data, si_weights)
}
\arguments{
\item{data}{Dataframe to be used for predicting reproductive number.}

\item{si_weights}{Matrix of probabilities of infection based on SI. Output
from \code{get_si_weights}.}
}
\value{
List of effective reproductive numbers based on the time list
  passed into \code{get_si_weights}.
}
\description{
Finds reproductive numbers based on number of individuals infected at each
  time point. Reweighs the \code{si_weights} by number infected.
}
\section{Warning}{

Data must contain 'cases' column.
}

