% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rts_setup.R
\name{get_si_weights}
\alias{get_si_weights}
\title{Serial Interval Based Probability of Infection}
\usage{
get_si_weights(t_list, ps_si)
}
\arguments{
\item{t_list}{List of times. For weekly data, this list will be the week
number starting at 1.}

\item{ps_si}{List of discrete serial interval probabilities. Result of
\code{get_ps}.}
}
\value{
Matrix of probabilities of infection based solely on serial
  interval.
}
\description{
Gets probability of infection based solely on serial interval
  (ie. time between primary and secondary infection.)
}
