% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{read_cdc_data}
\alias{read_cdc_data}
\title{Read CDC Data}
\usage{
read_cdc_data(data_to_read, region = FALSE)
}
\arguments{
\item{data_to_read}{A string, path to data .csv file.}

\item{region}{\code{TRUE} if incorporating interaction between regions,
\code{FALSE} otherwise.}
}
\value{
Dataframe with cleaned up CDC data.
}
\description{
Reads in CDC data from file and puts data into correct format.
}
\section{Warning}{

Data must contain 'cases' column and have time as the first column.
}

\examples{
dat <- read_cdc_data("D:/School/TB_Research/data/cdc_cases.csv",
  region = FALSE)
}
